analyzer/requirements.txt
paho-mqtt
influxdb-client
pandas
numpy